--HG--
extra : convert_revision : 455096c324b224614bcdace253d5d19be9c14f3d

+++ pidgin/win32/winpidgin.c
@@ -467,7 +467,7 @@ static BOOL winpidgin_set_running(BOOL fail_if_running) {
 
 			return FALSE;
 		} else
-			printf("Error (%d) accessing \"pidgin_is_running\" mutex.\n", err);
+			printf("Error (%u) accessing \"pidgin_is_running\" mutex.\n", (UINT) err);
 	}
 	return TRUE;
 }
